+++ src/regexp_nfa.c
@@ -4800,7 +4800,7 @@ addstate_here(
 		emsg(_(e_maxmempat));
 		return NULL;
 	    }
-	    newl = (nfa_thread_T *)alloc(newsize);
+	    newl = (nfa_thread_T *)alloc((int)newsize);
 	    if (newl == NULL)
 		return NULL;
 	    l->len = newlen;
